Nightly update
Changed file name to PEP class name.
Converted user_input to use new class
